cambios en entidadjugador y logica, hitbox cambiar de float a int

+++ src/Entidades/EntidadLogica.java
@@ -3,7 +3,7 @@ package Entidades;
 public interface EntidadLogica {
 
 	public Skin getSkin();
-	public float getPosX();
-	public float getPosY();
+	public int getPosX();
+	public int getPosY();
 	
 }
